[Tema 3] Finalizare si refactorizare

+++ Teme/Tema3/filters.h
@@ -15,6 +15,6 @@ typedef int FILTER_STATUS;
 *   Pune filtrul unitate daca nu se ofera un tip de filtru valid.
 * 	Memoria pentru filtru trebuie sa fie deja alocata si sa fie de 9 floaturi.
 */
-FILTER_STATUS getFilter(float* filter, char* type);
+FILTER_STATUS getFilter(double* filter, char* type);
 
 #endif  /* !FILTERS_H */
